Working version
With automatic mode implemented that drives forward until the photo sensor (newly implemented) detects a light source.
The bluetooth modules communicate through single bytes and a lot of if-statements, as sending packaged two-byte data resulted in a large delay on the receiving end. Will perhaps debug and implement this later, as it would shorten and simplify the code quite a bit.